converted companies logic to use async instead of promise for cleaner code
